feat: 전체 사용자는 선택한 사진들에 대한 내용을 작성하여 업로드한다.

+++ src/main/java/com/iiie/server/repository/GalleryRepository.java
@@ -2,6 +2,5 @@ package com.iiie.server.repository;
 
 import com.iiie.server.domain.Gallery;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
 
-public interface GalleryRepository extends JpaRepository<Gallery, Integer> {}
+public interface GalleryRepository extends JpaRepository<Gallery, Long> {}
